Some changes

+++ src/main/java/com/minicare/model/Sitter.java
@@ -1,31 +1,31 @@
 package com.minicare.model;
 
 public class Sitter extends Member {
-    private int experience;
-    private float expectedPay;
+    private double experience;
+    private double expectedPay;
 
     public Sitter() {}
 
     public Sitter(String firstName, String lastName, String email, String phoneNumber, String address,
-                  String password, int experience, float expectedPay) {
+                  String password, double experience, double expectedPay) {
         super(firstName, lastName, email, phoneNumber, address, password, MemberType.SITTER);
         this.experience = experience;
         this.expectedPay = expectedPay;
     }
 
-    public int getExperience() {
+    public double getExperience() {
         return experience;
     }
 
-    public void setExperience(int experience) {
+    public void setExperience(double experience) {
         this.experience = experience;
     }
 
-    public float getExpectedPay() {
+    public double getExpectedPay() {
         return expectedPay;
     }
 
-    public void setExpectedPay(float expectedPay) {
+    public void setExpectedPay(double expectedPay) {
         this.expectedPay = expectedPay;
     }
 }
